requirements.txt
fastapi
uvicorn[standard]
python-multipart

langgraph
langgraph-checkpoint-sqlite

openai
python-dotenv

pandas
numpy
openpyxl
scipy
simpy
matplotlib
xarray
python-dateutil

chromadb
sentence-transformers
pypdf
python-docx
tqdm

psycopg[binary]